pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clawshield"
version = "0.1.0"
description = "ClawShield bootstrap backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "pydantic-settings>=2.3.0,<3.0.0",
  "sqlalchemy>=2.0.30,<3.0.0",
  "PyYAML>=6.0.1,<7.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
